Share mocks between regular and equivalence fixtures
