+++ scripts/fiber/tests-failing.txt
@@ -1,3 +1,1 @@
-src/isomorphic/modern/class/__tests__/ReactClassEquivalence-test.js
-* tests the same thing for es6 classes and CoffeeScript
-* tests the same thing for es6 classes and TypeScript
+
