+++ scripts/fiber/tests-passing.txt
@@ -194,6 +194,10 @@ src/isomorphic/classic/element/__tests__/ReactElementValidator-test.js
 * does not blow up with inlined children
 * does not blow up on key warning with undefined type
 
+src/isomorphic/modern/class/__tests__/ReactClassEquivalence-test.js
+* tests the same thing for es6 classes and CoffeeScript
+* tests the same thing for es6 classes and TypeScript
+
 src/isomorphic/modern/class/__tests__/ReactCoffeeScriptClass-test.coffee
 * preserves the name of the class for use in error messages
 * throws if no render function is defined
