+++ scripts/jest/setupSpecEquivalenceReporter.js
@@ -9,6 +9,8 @@
 
 'use strict';
 
+require('./setupMocks');
+
 var expect = global.expect;
 
 var numExpectations = 0;
